requirements.txt
torch>=2.0
numpy>=1.24
tqdm>=4.66
scikit-learn>=1.3
pyyaml>=6.0
